net/kparser: Add kParser KMOD in net
kParser stands for "The Kernel Parser". This is a programmable network packet
parser which is a ported version of the PANDA parser.

For the introduction to the basic building blocks of PANDA parser, kindly refer
https://github.com/panda-net/panda/blob/main/documentation/parser.md

The iproute2 CLI for the kParser works in tandem with the KMOD kParser. Hence
the KMOD kParser must be loaded before the kParser enabled iproute2 CLI can be
used to configure and use it.

The kParser is hightly programmable, to add any types of network protocol packet
parsing support, there is no need to write or compile new code. kParser can be
configured to parse any networking protocol packet. Only the kParser CLI is
sufficient to configure and teach the KMOD kParser about how to parse any new
protocol and add that capability to Linux Kernel.

This can be configured via iproute2 kParser CLI

For the user space cli, please refer kParser changes in iproute2 project.

Signed-off-by: Pratyush <pratyush@sipanda.io>